refactor: extract workflow timeout race and park merge policy on MergeStrategy
Keep Workflow.swift under the 1k-line boundary and stop switching on merge
strategy in three places. Behavior is unchanged.
